+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Changelog
 
 ### Fixes
 
+- [IR] Always use a provider field if multiple provider and scalar refs are found. Previously we would possibly use just a scalar getter field wrapped in `InstanceFactory` for provider refs.
+
 ### Changes
 
 ### Contributors